pyproject.toml
[project]
name = "pycharts_timescaledb"
version = "0.0.0"
description = "Postgres + TimescaleDB Extention for use with Lightweight-Pycharts' Broker-API Plugin System."
readme = { file = "README.md", content-type = "text/markdown" }
authors = [{ name = "jack_of_some_trades" }]
license = { text = "TBD" }
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.2",
    "pandas-market-calendars>=5.0.0",
    "psycopg[binary,pool]>=3.2.6",
    "python-dotenv>=1.0.1",
]

[tool.setuptools]
packages = ["pycharts_timescaledb"]

[dependency-groups]
dev = [
    "alpaca-py>=0.39.4",
]
